Added Sending mail through Excel directly using outlook
Removing the otherwise "only" manual part of sending the excel reports. Using a  click of button now, the excel will attach itself to outlook and will be sent to the desired address. The option sheet consists of destination address, cc list, bcc list, subject and content. It also has an option to select the sheets of the workbook to be mailed.